+++ extensions/remix-icon/src/IconActionPanel.tsx
@@ -1,5 +1,5 @@
 import { ActionPanel, Action } from "@raycast/api";
-import { getSvgContent, svgToDataUri, toReactComponentName } from "./utils";
+import { getSvgContent, svgToDataUri, toComponentName } from "./utils";
 import metadata from "../assets/metadata.json";
 
 export default function IconActionPanel({
@@ -11,7 +11,7 @@ export default function IconActionPanel({
   iconName: string;
   updateRecentIcons: (category: string, iconName: string) => void;
 }>) {
-  const reactComponentName = toReactComponentName(iconName);
+  const componentName = toComponentName(iconName);
   const cdnLink = `<link href="https://cdn.jsdelivr.net/npm/remixicon@${metadata.version}/fonts/remixicon.css" rel="stylesheet"/>`;
 
   return (
@@ -22,16 +22,49 @@ export default function IconActionPanel({
         onCopy={() => updateRecentIcons(category, iconName)}
       />
       <Action.CopyToClipboard
-        title="Copy React Component"
-        content={`<${reactComponentName} size={24} color="currentColor" />`}
-        onCopy={() => updateRecentIcons(category, iconName)}
-      />
-      <Action.CopyToClipboard
         title="Copy Data URI"
         content={svgToDataUri(getSvgContent(category, iconName))}
         shortcut={{ modifiers: ["cmd"], key: "u" }}
         onCopy={() => updateRecentIcons(category, iconName)}
       />
+      <ActionPanel.Section title="React">
+        <Action.CopyToClipboard
+          title="Copy React Component"
+          content={`<${componentName} size={24} color="currentColor" />`}
+          shortcut={{ modifiers: ["cmd"], key: "r" }}
+          onCopy={() => updateRecentIcons(category, iconName)}
+        />
+        <Action.CopyToClipboard
+          title="Copy React Import"
+          content={`import { ${componentName} } from "@remixicon/react";`}
+          onCopy={() => updateRecentIcons(category, iconName)}
+        />
+        <Action.CopyToClipboard
+          // eslint-disable-next-line @raycast/prefer-title-case
+          title="Copy React NPM Install"
+          content={`npm install @remixicon/react`}
+          onCopy={() => updateRecentIcons(category, iconName)}
+        />
+      </ActionPanel.Section>
+      <ActionPanel.Section title="Vue">
+        <Action.CopyToClipboard
+          title="Copy Vue Component"
+          content={`<${componentName} size="24px" color="currentColor" />`}
+          shortcut={{ modifiers: ["cmd"], key: "v" }}
+          onCopy={() => updateRecentIcons(category, iconName)}
+        />
+        <Action.CopyToClipboard
+          title="Copy Vue Import"
+          content={`import { ${componentName} } from "@remixicon/vue";`}
+          onCopy={() => updateRecentIcons(category, iconName)}
+        />
+        <Action.CopyToClipboard
+          // eslint-disable-next-line @raycast/prefer-title-case
+          title="Copy Vue NPM Install"
+          content={`npm install @remixicon/vue`}
+          onCopy={() => updateRecentIcons(category, iconName)}
+        />
+      </ActionPanel.Section>
       <ActionPanel.Section title="Webfont">
         <Action.CopyToClipboard
           title="Copy HTML Tag"
@@ -47,7 +80,7 @@ export default function IconActionPanel({
         />
         <Action.CopyToClipboard
           // eslint-disable-next-line @raycast/prefer-title-case
-          title="Copy NPM Install Command"
+          title="Copy Webfont NPM Install"
           content={`npm install remixicon --save`}
           shortcut={{ modifiers: ["cmd"], key: "n" }}
           onCopy={() => updateRecentIcons(category, iconName)}